Patches to nlp_refactoring. (#493)
* NLP refactoring patches.


* Fixed a bug.


* Fixed a bug.


* Fixed a bug.


* Fixed a bug.


* Fixed a bug.


* Fixed a bug.


* Fixed a bug.


* Fixed a bug.


* Fixed a bug.


* Fixed a bug.


* Fixed a bug.


* U


* Docs updated.


* Added do_lower_case to all dataset import scripts.


* Fixed shuffle and bert_model_name.


* Fixed lower_case bug.


* Added support for other BERT models.


* fixed bug.


* fixed bug.


* Updated inference examples.


* Fixed jenkins.


* Fixed bug.


* Fixed bug.


* Fixed bug.


* Fixed bug.


* Cleaned up scripts.


* fixed style.


* fixed style.


* updated chinese docs.


* updated chinese docs.


* Updated style.
