+++ docs/sources/source/nlp/asr-improvement.rst
@@ -1,5 +1,5 @@
 Tutorial
-===========================
+========
 
 In this tutorial we will train an ASR postprocessing model to correct mistakes in
 output of end-to-end speech recognition model. This model method works similar to translation model in contrast to traditional ASR language model rescoring.
