+++ docs/sources/source/nlp/bert_pretraining.rst
@@ -1,5 +1,5 @@
-Pretraining BERT
-================
+Tutorial
+========
 
 In this tutorial, we will build and train a masked language model, either from scratch or from a pretrained BERT model, using the BERT architecture :cite:`nlp-bert-devlin2018bert`.
 Make sure you have ``nemo`` and ``nemo_nlp`` installed before starting this tutorial. See the :ref:`installation` section for more details.
