+++ nemo/collections/nlp/data/datasets/joint_intent_slot_dataset/__init__.py
@@ -14,6 +14,6 @@
 # limitations under the License.
 # =============================================================================
 
-from nemo.collections.nlp.data.datasets.joint_intent_slot_dataset.data_descriptor import *
 from nemo.collections.nlp.data.datasets.joint_intent_slot_dataset.inference_utils import *
 from nemo.collections.nlp.data.datasets.joint_intent_slot_dataset.joint_intent_slot_dataset import *
+from nemo.collections.nlp.data.datasets.joint_intent_slot_dataset.joint_intent_slot_descriptor import *
