+++ nemo/collections/nlp/data/datasets/text_classification/text_classification_dataset.py
@@ -48,7 +48,16 @@ class BertTextClassificationDataset(Dataset):
             If -1, use all dataset. Useful for testing.
     """
 
-    def __init__(self, input_file, max_seq_length, tokenizer, num_samples=-1, shuffle=False, use_cache=False):
+    def __init__(
+        self,
+        input_file,
+        max_seq_length,
+        tokenizer,
+        num_samples=-1,
+        shuffle=False,
+        use_cache=False,
+        do_lower_case=False,
+    ):
         self.input_file = input_file
         self.max_seq_length = max_seq_length
         self.tokenizer = tokenizer
@@ -56,6 +65,7 @@ class BertTextClassificationDataset(Dataset):
         self.use_cache = use_cache
         self.shuffle = shuffle
         self.vocab_size = self.tokenizer.tokenizer.vocab_size
+        self.do_lower_case = do_lower_case
 
         if use_cache:
             data_dir, filename = os.path.split(input_file)
@@ -70,7 +80,6 @@ class BertTextClassificationDataset(Dataset):
 
         if use_cache and os.path.exists(cached_features_file):
             self.load_cached_features(cached_features_file)
-
         else:
             with open(input_file, "r") as f:
                 sent_labels, all_sent_subtokens = [], []
@@ -90,9 +99,13 @@ class BertTextClassificationDataset(Dataset):
                     if index % 20000 == 0:
                         logging.debug(f"Processing line {index}/{len(lines)}")
 
-                    sent_label = int(line.split()[-1])
+                    if do_lower_case:
+                        line = line.lower()
+
+                    line_splited = line.strip().split()
+                    sent_label = int(line_splited[-1])
                     sent_labels.append(sent_label)
-                    sent_words = line.strip().split()[:-1]
+                    sent_words = line_splited[:-1]
                     sent_subtokens = [tokenizer.cls_token]
 
                     for word in sent_words:
